refactor: improve C# SDK examples structure
- Refactor examples to use classes instead of top-level statements
- Add Program.cs entry point with command routing (main/upload)
- Fix type cast for thread.Id (long to int)
- Fix System.IO namespace conflicts in Upload example
- Add bin/, obj/ to gitignore
